% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county2010_to_puma2012_pop14}
\alias{county2010_to_puma2012_pop14}
\title{County10-PUMA12 Crosswalk File, with 2014 population as weights}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4546 rows and 11 columns.}
\source{
\url{http://mcdc.missouri.edu/websas/geocorr14.html}
}
\usage{
county2010_to_puma2012_pop14
}
\description{
County10-PUMA12 Crosswalk File, with 2014 population as weights
}
\keyword{datasets}
