% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county2014_to_puma2000}
\alias{county2014_to_puma2000}
\title{County14-PUMA00 Crosswalk File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4432 rows and 11 columns.}
\source{
\url{http://mcdc.missouri.edu/websas/geocorr14.html}
}
\usage{
county2014_to_puma2000
}
\description{
County14-PUMA00 Crosswalk File
}
\keyword{datasets}
