% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puma2000_to_puma2012}
\alias{puma2000_to_puma2012}
\title{PUMA00-PUMA12 Crosswalk File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4444 rows and 11 columns.}
\source{
\url{http://mcdc.missouri.edu/websas/geocorr14.html}
}
\usage{
puma2000_to_puma2012
}
\description{
PUMA00-PUMA12 Crosswalk File
}
\keyword{datasets}
