% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta2010_to_puma2012}
\alias{zcta2010_to_puma2012}
\title{ZCTA10-PUMA12 Crosswalk File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44493 rows and 11 columns.}
\source{
\url{http://mcdc.missouri.edu/websas/geocorr14.html}
}
\usage{
zcta2010_to_puma2012
}
\description{
ZCTA10-PUMA12 Crosswalk File
}
\keyword{datasets}

