% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta2010_to_usd2014}
\alias{zcta2010_to_usd2014}
\title{ZCTA10-USD14 Crosswalk File}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67592 rows and 11 columns.}
\source{
\url{http://mcdc.missouri.edu/websas/geocorr14.html}
}
\usage{
zcta2010_to_usd2014
}
\description{
ZCTA10-USD14 Crosswalk File
}
\keyword{datasets}
